Styling for detail page + new buttons for deleting and editing

+++ app/src/main/java/com/jbol/dailydrops/DetailsActivity.java
@@ -5,10 +5,14 @@ import android.graphics.Bitmap;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+import android.widget.ImageButton;
 import android.widget.ImageView;
+import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
+
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.constraintlayout.widget.ConstraintLayout;
 import com.jbol.dailydrops.database.SQLiteDataBaseHelper;
 import com.jbol.dailydrops.models.GlobalDropModel;
 import com.jbol.dailydrops.services.DateService;
@@ -17,9 +21,10 @@ import com.jbol.dailydrops.services.AsyncURLService;
 import java.time.format.FormatStyle;
 
 public class DetailsActivity extends AppCompatActivity {
-    private TextView tv_title, tv_date;
-    private Button btn_delete, btn_edit, btn_back;
-    private ImageView iv_image;
+    private TextView tv_title, tv_date, tv_note;
+    private ImageButton ib_delete, ib_edit;
+    private ImageView iv_image, iv_back_btn;
+    private RelativeLayout cl_root;
 
     private GlobalDropModel drop;
 
@@ -31,20 +36,33 @@ public class DetailsActivity extends AppCompatActivity {
         Intent intent = getIntent();
         drop = (GlobalDropModel) intent.getSerializableExtra("drop");
 
+        cl_root = findViewById(R.id.cl_root);
+
+        if (drop.getType().equals(GlobalDropModel.ONLINE_TYPE)) {
+            cl_root.setBackgroundColor(getResources().getColor(R.color.colorOnline));
+        } else {
+            cl_root.setBackgroundColor(getResources().getColor(R.color.colorLocal));
+        }
+
         tv_title = findViewById(R.id.tv_title);
         tv_title.setText(drop.getTitle());
 
-        initializeImage();
+        tv_note = findViewById(R.id.tv_note);
+        tv_note.setText(drop.getNote());
+
         initializeBackBtn();
         initializeDate();
         initializeDeleteBtn();
         initializeEditBtn();
+        initializeImage();
     }
 
     private void initializeImage() {
         iv_image = findViewById(R.id.iv_image);
 
         if (drop.getImage() == null) {
+            iv_image.setVisibility(View.INVISIBLE);
+            iv_image.setMaxHeight(100);
             return;
         }
 
@@ -64,38 +82,42 @@ public class DetailsActivity extends AppCompatActivity {
 
     // Can only be executed if it's a local drop
     private void initializeDeleteBtn() {
-        btn_delete = findViewById(R.id.btn_delete);
+        ib_delete = findViewById(R.id.ib_delete);
 
         if (drop.getType().equals(GlobalDropModel.ONLINE_TYPE)) {
-            btn_delete.setVisibility(View.INVISIBLE);
+            ib_delete.setVisibility(View.GONE);
             return;
         }
-        btn_delete.setOnClickListener(v -> {
+        ib_delete.setOnClickListener(v -> {
             SQLiteDataBaseHelper sqldbHelper = SQLiteDataBaseHelper.getHelper(DetailsActivity.this);
             boolean success = sqldbHelper.deleteDrop(Integer.parseInt(drop.getId()));
+            if (!success) {
+                Toast.makeText(this, "Drop couldn't be deleted. Please try again.", Toast.LENGTH_SHORT).show();
+                return;
+            }
             ImageService.deleteImageFromStorage(this, Integer.parseInt(drop.getId()));
-
-            Toast.makeText(DetailsActivity.this, "Success= " + success, Toast.LENGTH_SHORT).show();
+            Intent i = new Intent(DetailsActivity.this, MainActivity.class);
+            DetailsActivity.this.startActivity(i);
         });
     }
 
     private void initializeBackBtn() {
-        btn_back = findViewById(R.id.btn_back);
+        iv_back_btn = findViewById(R.id.iv_back_btn);
 
-        btn_back.setOnClickListener(v -> {
+        iv_back_btn.setOnClickListener(v -> {
             Intent i = new Intent(DetailsActivity.this, MainActivity.class);
             DetailsActivity.this.startActivity(i);
         });
     }
 
     private void initializeEditBtn() {
-        btn_edit = findViewById(R.id.btn_edit);
+        ib_edit = findViewById(R.id.ib_edit);
 
         if (drop.getType().equals(GlobalDropModel.ONLINE_TYPE)) {
-            btn_edit.setVisibility(View.INVISIBLE);
+            ib_edit.setVisibility(View.GONE);
             return;
         }
-        btn_edit.setOnClickListener(v -> {
+        ib_edit.setOnClickListener(v -> {
             Intent i = new Intent(DetailsActivity.this, EditActivity.class);
             i.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
             i.putExtra("drop", drop);
